requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0.0
langchain>=0.3.0
langchain-chroma>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-multipart
pyyaml>=6.0
langchain-huggingface